requirements.txt
stix2
sentence-transformers
tqdm
qdrant-client
google-generativeai